Diverse amélioration : - Correction TU - Déplacement des Validateurs et annotations dans un package indépendants - Passage de service en interface avec Implémentation et appel de ces interfaces dans les couches supérieurs. - MAJ Javadoc.
TODO :
- Finir le transfert des services vers interfaces
- Finir javadoc
- Améliorer TU avec ajout des tests
