+++ src/main/java/com/axoulotl/alextheque/controller/GameController.java
@@ -7,6 +7,7 @@ import com.axoulotl.alextheque.model.dto.input.GameUpdateDTO;
 import com.axoulotl.alextheque.model.dto.output.ErrorDTO;
 import com.axoulotl.alextheque.model.dto.output.GameOutputDTO;
 import com.axoulotl.alextheque.service.GameService;
+import com.axoulotl.alextheque.service.GameServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -14,6 +15,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.PositiveOrZero;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -22,14 +24,10 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1")
+@RequiredArgsConstructor
 public class GameController {
 
-    GameService gameService;
-
-    @Autowired
-    public GameController(GameService gameService) {
-        this.gameService = gameService;
-    }
+    GameService gameServiceImpl;
 
     @Operation(summary = "Add a new game in collection")
     @ApiResponses(value = {
@@ -53,7 +51,7 @@ public class GameController {
     public ResponseEntity<Object> addGame(@Valid @RequestBody GameDTO gameDTO) {
         ResponseEntity<Object> responseEntity = null;
         try {
-            responseEntity = ResponseEntity.ok().body(gameService.addGame(gameDTO));
+            responseEntity = ResponseEntity.ok().body(gameServiceImpl.addGame(gameDTO));
         } catch (AlexthequeStandardError ex) {
             if (ex.getError() == StandardErrorEnum.ERROR_INPUT)
                 responseEntity = ResponseEntity.badRequest().body(new ErrorDTO(ex.getComment(), ex.getError()));
@@ -88,7 +86,7 @@ public class GameController {
         ResponseEntity<Object> responseEntity;
 
         try {
-            responseEntity = ResponseEntity.ok().body(gameService.getAllGames(page, size));
+            responseEntity = ResponseEntity.ok().body(gameServiceImpl.getAllGames(page, size));
         } catch (AlexthequeStandardError ex) {
             responseEntity = ResponseEntity.badRequest().body(new ErrorDTO(ex.getComment(), ex.getError()));
         } catch (Exception ex) {
@@ -119,7 +117,7 @@ public class GameController {
     public ResponseEntity<Object> getGameById(@PositiveOrZero @PathVariable Integer id) {
         ResponseEntity<Object> responseEntity;
         try {
-            responseEntity = ResponseEntity.ok().body(gameService.getGameFromId(id));
+            responseEntity = ResponseEntity.ok().body(gameServiceImpl.getGameFromId(id));
         } catch (AlexthequeStandardError ex) {
             responseEntity = ResponseEntity.internalServerError().body(new ErrorDTO(ex.getComment(), ex.getError()));
         }
@@ -150,7 +148,7 @@ public class GameController {
             @Valid @RequestBody GameUpdateDTO gameUpdateDTO) {
         ResponseEntity<Object> responseEntity = null;
         try {
-            responseEntity = ResponseEntity.ok().body(gameService.updateGameFromId(id, gameUpdateDTO));
+            responseEntity = ResponseEntity.ok().body(gameServiceImpl.updateGameFromId(id, gameUpdateDTO));
         } catch (AlexthequeStandardError ex) {
             if (ex.getError() == StandardErrorEnum.ERROR_INPUT)
                 responseEntity = ResponseEntity.badRequest().body(new ErrorDTO(ex.getComment(), ex.getError()));
