+++ src/main/java/com/axoulotl/alextheque/service/SearchServiceImpl.java
@@ -5,12 +5,10 @@ import com.axoulotl.alextheque.exception.StandardErrorEnum;
 import com.axoulotl.alextheque.model.dto.input.SearchGameDTO;
 import com.axoulotl.alextheque.model.dto.output.GamesOutputDTO;
 import com.axoulotl.alextheque.model.entity.Game;
-import com.axoulotl.alextheque.model.entity.enums.Status;
 import com.axoulotl.alextheque.repository.GameRepository;
 import com.axoulotl.alextheque.repository.specification.GameSpecifications;
 import com.axoulotl.alextheque.service.converter.GameToGameDTOConverter;
 import jakarta.persistence.EntityManager;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -19,15 +17,15 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 @Service
-public class SearchService {
+public class SearchServiceImpl {
 
     GameRepository gameRepository;
     GameToGameDTOConverter converter;
     EntityManager entityManager;
 
     @Autowired
-    public SearchService(GameRepository gameRepository, GameToGameDTOConverter converter,
-                         EntityManager entityManager) {
+    public SearchServiceImpl(GameRepository gameRepository, GameToGameDTOConverter converter,
+                             EntityManager entityManager) {
         this.gameRepository = gameRepository;
         this.converter = converter;
         this.entityManager = entityManager;
