+++ src/main/java/com/axoulotl/alextheque/validation/annotation/ValidDateRange.java
@@ -1,5 +1,6 @@
-package com.axoulotl.alextheque.service.validation;
+package com.axoulotl.alextheque.validation.annotation;
 
+import com.axoulotl.alextheque.validation.validator.DateRangeValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
