+++ src/main/java/com/axoulotl/alextheque/validation/validator/DateRangeValidator.java
@@ -1,8 +1,8 @@
-package com.axoulotl.alextheque.service.validation;
+package com.axoulotl.alextheque.validation.validator;
 
+import com.axoulotl.alextheque.validation.annotation.ValidDateRange;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
-import org.springframework.beans.BeanWrapper;
 import org.springframework.beans.BeanWrapperImpl;
 
 import java.time.LocalDate;
