+++ src/test/java/com/axoulotl/alextheque/services/GameServiceImplTest.java
@@ -31,7 +31,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @ExtendWith(MockitoExtension.class)
-public class GameServiceTest {
+public class GameServiceImplTest {
 
     @Mock
     GameRepository gameRepository;
@@ -43,7 +43,7 @@ public class GameServiceTest {
     GameToGameDTOConverter converter;
 
     @InjectMocks
-    GameService gameService;
+    GameService gameServiceImpl;
 
     @Test
     void whenAddGame_GivenGoodDto_TheGameIsSaved() throws AlexthequeStandardError {
@@ -58,7 +58,7 @@ public class GameServiceTest {
         when(gameRepository.save(any(Game.class))).thenReturn(game);
         when(converter.gameToGameDTO(any(Game.class))).thenCallRealMethod();
 
-        GameOutputDTO result = gameService.addGame(gameDTO);
+        GameOutputDTO result = gameServiceImpl.addGame(gameDTO);
 
         assertThat(result).isNotNull();
         assertThat(result.getId()).isEqualTo(1);
@@ -75,7 +75,7 @@ public class GameServiceTest {
         when(consoleRepository.findById(any(Integer.class))).thenReturn(console);
         doThrow(new RuntimeException("Simulated DB Connection error")).when(gameRepository).save(any(Game.class));
 
-        AlexthequeStandardError thrownException = assertThrows(AlexthequeStandardError.class, () -> gameService.addGame(gameDTO));
+        AlexthequeStandardError thrownException = assertThrows(AlexthequeStandardError.class, () -> gameServiceImpl.addGame(gameDTO));
 
         verify(gameRepository, times(1)).save(any(Game.class));
         assertThat(thrownException.getError()).isEqualTo(StandardErrorEnum.ERROR_DATABASE);
@@ -104,7 +104,7 @@ public class GameServiceTest {
         List<GameOutputDTO> gameOutputDTOList = Collections.singletonList(gameOutputDTO);
         when(converter.gamesToListOfGames(anyList())).thenReturn(gameOutputDTOList);
 
-        GamesOutputDTO result = gameService.getAllGames(page, size);
+        GamesOutputDTO result = gameServiceImpl.getAllGames(page, size);
 
         assertThat(result).isNotNull();
         assertThat(result.getNbGames()).isEqualTo(11L);
