+++ src/test/java/com/axoulotl/alextheque/services/SearchServiceImplTest.java
@@ -5,7 +5,7 @@ import com.axoulotl.alextheque.model.dto.input.SearchGameDTO;
 import com.axoulotl.alextheque.model.dto.output.GamesOutputDTO;
 import com.axoulotl.alextheque.model.entity.Game;
 import com.axoulotl.alextheque.repository.GameRepository;
-import com.axoulotl.alextheque.service.SearchService;
+import com.axoulotl.alextheque.service.SearchServiceImpl;
 import com.axoulotl.alextheque.service.converter.GameToGameDTOConverter;
 import com.axoulotl.alextheque.services.utils.UtilsTest;
 import org.junit.jupiter.api.BeforeEach;
@@ -27,10 +27,10 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-public class SearchServiceTest {
+public class SearchServiceImplTest {
 
     @InjectMocks
-    private SearchService searchService;
+    private SearchServiceImpl searchServiceImpl;
 
     @Mock
     private GameRepository gameRepository;
@@ -61,7 +61,7 @@ public class SearchServiceTest {
         when(converter.gamesToListOfGames(anyList()))
                 .thenCallRealMethod();
 
-        GamesOutputDTO result = searchService.searchGameWithCriteria(searchGameDTO, page, size);
+        GamesOutputDTO result = searchServiceImpl.searchGameWithCriteria(searchGameDTO, page, size);
 
         assertNotNull(result);
         assertEquals(1, result.getNbGames());
